refactor: extract dataset fetching functions into separate files
Separates each dataset fetching function into its own file
for better organization and maintainability.
